[Refactor] ShellTestAppTest.cpp > test.cpp, test.h
